test/integ.repo-patrol.js.snapshot/asset.4973d8f566643ce42a69ab2c3539617d5f093d50ea8021b868d924732f79633e/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0
boto3>=1.35.0
PyGithub>=2.5.0
cryptography>=44.0.0
